+++ tests/test_windfarm.py
@@ -21,7 +21,6 @@ from tests.conftest import (
 )
 
 
-# @pytest.mark.cat("all")
 def test_windfarm_init(env_setup):
     """Tests the setup of ``Windfarm``."""
     env = env_setup
@@ -211,7 +210,7 @@ def test_windfarm_init(env_setup):
         assert cable.env == correct_cable.env
         assert cable.id == correct_cable.id
         assert cable.name == correct_cable.name
-        assert cable.turbine == correct_cable.turbine
+        assert cable.system == correct_cable.system
         assert cable.data.name == cable.data.name
         assert cable.operating_level == correct_cable.operating_level == 1
         assert cable.broken == correct_cable.broken is False
@@ -221,7 +220,6 @@ def test_windfarm_init(env_setup):
         assert cable.processes.keys() == correct_cable.processes.keys()
 
 
-# @pytest.mark.cat("all")
 def test_windfarm_failed_init(env_setup):
     """Tests the failing cases for setup of ``Windfarm`` where no data is provided."""
     env = env_setup